chore: simplify NotificationProductaanvraagTest by removing redundant CMMN flow scenarios and renaming to NotificationProductaanvraagBpmnTest
